Fill Movements list from the API, instead of hard-coded in the app itself.
Some functionalities like sorting, pagination and filtering were lost in the process, should fix them briefly
The unit test of MovementsListComponent is failing due some problems in the MovementsService. It would be good to ensure that all unit tests are passing, should try to solve this briefly too.

A good next step after solving the above-mentioned problems would be to containarize this app and add it to the docker-compose that alreadycontains the database and the backend.